Rename handlers.py to config.py
