+++ mlflow/gateway/app.py
@@ -6,7 +6,7 @@ from traceback import format_exc
 from typing import List, Union
 
 from mlflow.version import VERSION
-from mlflow.gateway.handlers import Route, RouteConfig, _route_config_to_route, _load_route_config
+from mlflow.gateway.config import Route, RouteConfig, _route_config_to_route, _load_route_config
 
 
 _logger = logging.getLogger(__name__)
